qcacld-3.0: update sap interface enum check to bit mask
This change is to update sap interface restriction mask
enum check to NL80211_IFTYPE_AP bit mask.

CRs-Fixed: 3118953

+++ core/hdd/inc/wlan_hdd_main.h
@@ -2062,7 +2062,7 @@ struct hdd_context {
 	uint16_t unsafe_channel_list[NUM_CHANNELS];
 #endif /* FEATURE_WLAN_CH_AVOID */
 #ifdef FEATURE_WLAN_CH_AVOID_EXT
-	uint8_t restriction_mask;
+	uint32_t restriction_mask;
 #endif
 
 	uint8_t max_intf_count;
